schema for private user data form
